+++ cpp/testsuite/ApplicationCases/RenderparkApplication/src/options/OptionsGroupToneMapping.h
@@ -28,7 +28,7 @@ class OptionsGroupToneMapping final {
     static void blueChromaOption(Vector3D &value);
     static void whiteChromaOption(Vector3D &value);
     static void toneMappingCommandLineOptionDescAdaptMethodOption(char *&value);
-    static void gammaOption(float &value);
+    static void gammaOption(double &value);
     static bool parseColor3(int argc, char **argv, ColorRgb &value);
     static bool parseCieXy(int argc, char **argv, Vector3D &value);
 };
